Add files via upload
Number Guessing Game :
The computer generates a random number, and the user has to guess it (Range - 1 to 100). The user will input their guess, and your program provides feedback on whether the guess is too high, too low, or correct along with keeping a track on the number of attempts.